Start rebuilding the data store communication
In light of the DAO pattern commonly seen in Java systems, rebuilding
the database communication to be a system of Data Sources and Data
Stores. Data Sources handle the raw communication with whereever the
data is stored. Data Stores handle the conversion between the Data
Source and the actual objects in question, like a Construct, Part, or
UserSettings.